% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacebucket-package.R
\docType{data}
\name{ABC}
\alias{ABC}
\alias{A}
\alias{B}
\alias{C}
\title{Simple data examples}
\description{
A, B, C are three partially overlaying polygon layers with
different sets of attributes.
}
