% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersections.R
\name{layer_n}
\alias{layer_n}
\title{N intersections}
\usage{
layer_n(x, n = 2, ..., keep_index = FALSE)
}
\arguments{
\item{x}{polymer}

\item{n}{minimum number of intersections to keep}

\item{...}{ignored for now}

\item{keep_index}{for expert use only, maintains the list of triangle indexes
on the sf output (and sf cannot plot if that is present)}
}
\value{
sf data frame
}
\description{
Find all fragments that are intersected by any other feature in any layer.
}
\details{
Returns a simple features data frame with all triangles that occur \code{n} times with
\code{n = 2} as a minimum. Each triangle feature contains a nested data frame in \code{idx}
that keeps the links to the input layers by \code{layer}, \code{object} and \code{path}.
}
\examples{
library(sf)
plot(A["layer"], reset = TRUE)
plot(B, add = TRUE, col = "hotpink")
plot(C, add = TRUE, col = "firebrick")

sb <- polymer(A, B, C)
plot(layer_n(sb), add = TRUE, col = "grey")
plot(layer_n(sb, n = 3), add = TRUE, col = "dodgerblue")
}
