% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersections.R
\name{n_intersections}
\alias{n_intersections}
\title{N intersections}
\usage{
n_intersections(x, n = 2, ...)
}
\arguments{
\item{x}{polymer}

\item{n}{minimum number of intersections to keep}

\item{...}{ignored for now}
}
\value{
sf data frame
}
\description{
Find all fragments that are intersected by any other feature in any layer.
}
\details{
Returns a simple features data frame with all triangles that occur \code{n} times with
\code{n = 2} as a minimum. Each triangle feature contains a nested data frame in \code{idx}
that keeps the links to the input layers by \code{layer}, \code{object} and \code{path}.
}
\examples{
plot(A["layer"], reset = TRUE)
plot(B, add = TRUE, col = "hotpink")
plot(C, add = TRUE, col = "firebrick")

sb <- polymer(A, B, C)
plot(n_intersections(sb), add = TRUE, col = "grey")
plot(n_intersections(sb, n = 3), add = TRUE, col = "dodgerblue")
}
