% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polymer.R
\name{plot.polymer}
\alias{plot.polymer}
\title{Plot the primitive space bucket}
\usage{
\method{plot}{polymer}(x, ...)
}
\arguments{
\item{x}{polymer}

\item{...}{arguments to \link{polypath}}
}
\value{
nothing
}
\description{
Plot the primitive space bucket
}
\examples{
plot(polymer(A, B, C))
library(sf)
example(st_read)
x <- nc[1:5, ]
bucket <- polymer(nc, st_jitter(nc, amount = 0.1))
plot(bucket)
}
