% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacebucket.R
\name{plot.spacebucket}
\alias{plot.spacebucket}
\title{Plot the primitive space bucket}
\usage{
\method{plot}{spacebucket}(x, ...)
}
\arguments{
\item{x}{spacebucket}

\item{...}{arguments to \link{polypath}}
}
\value{
nothing
}
\description{
Plot the primitive space bucket
}
\examples{
plot(spacebucket(A, B, C))
}
