% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polymer-package.R
\docType{package}
\name{polymer-package}
\alias{polymer-package}
\title{Polymer}
\description{
Find overlapping polygons regions by converting a collection of sf
data frame polygon layers to a single pool of triangles. The pool
is used for subsequent queries, what layers intersect and how, maintaining
links to input data of all layers.
}
