% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polymer.R
\name{print.polymer}
\alias{print.polymer}
\title{Print polymer}
\usage{
\method{print}{polymer}(x, ...)
}
\arguments{
\item{x}{polymer}

\item{...}{ignored}
}
\value{
x invisibly
}
\description{
Print a short description of the polymer contents.
}
\examples{
polymer(A, B, C)
}
