% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacebucket.R
\name{print.spacebucket}
\alias{print.spacebucket}
\title{Print spacebucket}
\usage{
\method{print}{spacebucket}(x, ...)
}
\arguments{
\item{x}{spacebucket}

\item{...}{ignored}
}
\value{
x invisibly
}
\description{
Print a short description of the bucket contents.
}
\examples{
spacebucket(A, B, C)
}
