% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacebucket-package.R
\docType{package}
\name{spacebucket-package}
\alias{spacebucket}
\alias{spacebucket-package}
\alias{spacebucket-package}
\title{spacebucket: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).

behold the spacebucket.
}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
