% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polymer.R
\name{spacebucket}
\alias{spacebucket}
\title{Space bucket}
\usage{
spacebucket(...)
}
\arguments{
\item{...}{sf polygon data frame inputs}
}
\value{
a spacebucket, see details
}
\description{
Convert a collection of sf data frame polygon layers to a single pool
of triangles.
}
\details{
Each triangle is identified by which path in the inputs it belongs to. None of
this is very useable yet. Holes can be identified but aren't at the moment, any
path that is a hole is identified per triangle.

\code{input} is a list with all input objects
\code{primitives} is the triangulation object
\code{geometry_map} is the paths with their row count
\code{index} is the mapping between triangle and path/s
}
\examples{
spacebucket(A, B, C)
}
